+++ src/main/java/com/nephest/battlenet/sc2/web/service/blizzard/BlizzardSC2API.java
@@ -131,7 +131,7 @@ public class BlizzardSC2API
         return client;
     }
 
-    public static BlizzardSeason getSeason(Long id)
+    public static BlizzardSeason getSeason(Integer id)
     {
         return MMR_SEASONS.get(id);
     }
